requirements.txt
# sudo apt-get -y install python3-rpi.gpio
# pip install RPi.GPIO 
# sudo pip3 install Adafruit_DHT

Adafruit-Blinka==7.1.0
adafruit-circuitpython-dht==3.7.1
Adafruit-PlatformDetect==3.20.1
Adafruit-PureIO==1.1.9
keyboard==0.13.5
pyftdi==0.53.3
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==4.3.2
RPi.GPIO==0.7.1
smbus==1.1.post2
sysv-ipc==1.1.0